Improve loading performance of assets (#17)
* [refs #00016] Remove @imported CSS

The @import directive creates a very synchronous, sequential way of
downloading CSS, completely removing our ability to parallelise. By
breaking our CSS out into two separate link elements in the HTML, we can
begin to download the files completely side-by-side.

* [refs #00016] Move script block above CSS

Inline scripts block on CSSOM construction, which means that they won’t
execute until all discovered CSS is downloaded and parsed. This means
that it’s actually better for performance to place these above any CSS
requests so as not to block them.

* [refs #00016] Move Twitter widget into head

Seeing as though we’re loading the Twitter widget.js asynchronously, it
doesn’t need to be defined late/JIT. We can move it up to the head to
get it coming over the wire ASAP, but still ensure that it’s not
blocking anything else.

* [refs #00016] Preconnect external origins

There are a few external domains that we hit on every page load; let’s
preemptively open up their connections.

* [refs #00016] Remove stray Gaug.es code

I didn’t realise the include for the Gaug.es analytics was included in
so many layouts. Oops! I’ve removed it so that we don’t have it doubled
up on so many pages.

* [refs #00016] Preconnect analytics origins

Along with all the other external domains we’ve preconnected, we might
as well do the analytics stuff, too.